:bug: Fixes Kong gateway annotations
After the deployment of our custom CRD trigger at QuintoAndar we've discovered a
minor bug with Gateway. For Kong gateway it is not being added the required
annotation for it to work.

On this commit I've added that annotation and also removed the rewrite-rule for Nginx.

Now, we can deploy Kong triggers.